src/requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
python-multipart
passlib
python-jose
